feat: allow opening nominate modal from multiple places (#220)
* feat: allow opening nominate modal from multiple places

* feat: fix build issues
