--- src/app/(pages)/(home)/@modal/(.)nominate/[walletId]/page.tsx
@@ -1,3 +1,0 @@
-import NominateBuilder from "@/app/_modals/nominate-builder/page";
-
-export default NominateBuilder;
