+++ src/app/(pages)/@modal/(...)nominate/[walletId]/loading.tsx
@@ -1,0 +1,3 @@
+import { default as Loading } from "@/app/(pages)/nominate/[walletId]/loading";
+
+export default Loading;
