+++ src/app/(pages)/@modal/(...)nominate/[walletId]/not-found.tsx
@@ -1,0 +1,3 @@
+import { default as NotFound } from "@/app/(pages)/nominate/[walletId]/not-found";
+
+export default NotFound;
