+++ src/app/(pages)/@modal/(...)nominate/[walletId]/page.tsx
@@ -1,0 +1,3 @@
+import { default as Page } from "@/app/(pages)/nominate/[walletId]/page";
+
+export default Page;
