+++ src/app/(pages)/@modal/tokenomics/page.tsx
@@ -1,0 +1,3 @@
+export default function Nothing() {
+  return null;
+}
