+++ src/app/(pages)/layout.tsx
@@ -1,0 +1,14 @@
+export default async function PagesLayout({
+  children,
+  modal,
+}: {
+  children: React.ReactNode;
+  modal: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+      {modal}
+    </>
+  );
+}
