+++ src/app/_api/functions/get-table-leaderboard-values.ts
@@ -4,22 +4,19 @@ import { unstable_cache } from "next/cache";
 import { supabase } from "@/db";
 import { TableLeaderboardValue } from "../../_components/table-leaderboard";
 import { getCurrentUser } from "../data/users";
-import { CACHE_5_MINUTES, CacheKey } from "../helpers/cache-keys";
+import { CacheKey } from "../helpers/cache-keys";
 
-const getLeaderboardTop50 = unstable_cache(
-  async () => {
-    const { data: leaderboardData } = await supabase
-      .from("boss_leaderboard")
-      .select("*, users(farcaster_id, passport_id)")
-      .order("rank", { ascending: true })
-      .order("passport_builder_score", { ascending: false })
-      .limit(50)
-      .throwOnError();
+const getLeaderboardTop50 = unstable_cache(async () => {
+  const { data: leaderboardData } = await supabase
+    .from("boss_leaderboard")
+    .select("*, users(farcaster_id, passport_id, last_wallet)")
+    .order("rank", { ascending: true })
+    .order("passport_builder_score", { ascending: false })
+    .limit(50)
+    .throwOnError();
 
-    return leaderboardData ?? [];
-  },
-  ["leaderboard_top_50" satisfies CacheKey],
-);
+  return leaderboardData ?? [];
+}, ["leaderboard_top_50" satisfies CacheKey]);
 
 export const getTableLeaderboardValues = async (): Promise<
   TableLeaderboardValue[]
@@ -31,7 +28,11 @@ export const getTableLeaderboardValues = async (): Promise<
   if (!containsUser && user?.boss_leaderboard) {
     leaderboard.push({
       ...user.boss_leaderboard,
-      users: { farcaster_id: null, passport_id: null },
+      users: {
+        farcaster_id: null,
+        passport_id: null,
+        last_wallet: user.wallet,
+      },
     });
   }
 
@@ -45,7 +46,8 @@ export const getTableLeaderboardValues = async (): Promise<
       bossScore: entry.boss_score,
       nominationsReceived: entry.nominations_received,
       rank: entry.rank?.toString() ?? "---",
-      farcaster_id: entry.users?.farcaster_id ?? undefined,
-      passport_id: entry.users?.passport_id ?? undefined,
+      farcasterId: entry.users?.farcaster_id ?? null,
+      passportId: entry.users?.passport_id ?? null,
+      walletAddress: entry.users?.last_wallet ?? null,
     }));
 };
